requirements.txt
streamlit
requests
google.generativeai